Implement filter and modify expressions
We haven't implemented parsing yet. The commit contains a usage
example, which will be removed when we implement the actual user
interface.

This commit adds two new functions:

- 'evaluate-modify-expression' takes a Racket-representation of a
modify expression. It returns a function that can be used to modify
some item in the item database.

- 'evaluate-filter-expression' takes a Racket-representation of a
filter expression. It returns a function that can be used to check
whether some item in the database matches that filter expression.

We also add the function 'search', which simply returns all items in
the database matching a particular filter expression.