Use hardcoded built-in property types everywhere
This commit removes a bunch of #:property-types parameters. We didn't use those
anyway, and in fact had implicit dependencies on the property-types argument
including a set of built-in properties. This was a poor solution that lead to a
lot of redundant code. For now, we'll replace it by just hardcoding the built-in
property types and depending on those everywhere.

In the future, we may still want to support some kind of customization of the
property types, but this feature is still quite far away, and it should probably
be implemented in a completely different way.