Implement date type and add placeholder "wait" property
We haven't implemented any functionality for the wait property yet, and we have
not even started implementing "date".

Also, the parser and representation for the date type are quite fragile. We
should probably use some actual date type in Racket to represent these, at least
when we want to do math on them. Also, we should verify somewhere that dates are
valid dates -- for instance, 2019-02-31 should probably display an error
somewhere.