Get rid of "expression" module
Actually, it may not be a bad idea to have a module like this, but the functions
in this module were essentially useless.

Also, this module was used very inconsistently anyway, so if we wanted to change
the representation of expressions, we would have to do it throughout the entire
codebase no matter what.