Add "+" operator to dates
This required re-doing how operators are evaluated in order to better deal with
optional types. The procedure evaluate-operator-expression must now be
explicitly passed the expected object type, which is used to find an appropriate
operator. (Previously, it would calculate the type based on the value of the
given object, but when the object is #f, it cannot correctly guess that the
object type is, say, '(opt date).)