Add initial work on links
This adds the "children" and "parents" types, which makes it possible to link
items together without introducing dependencies between them.

The children of an item is shown in the "tree" view, which has been improved to
avoid printing the same subtree multiple times.

In the future, we probably want to distinguish between children and dependencies
in the output. We also want a version of "tree" that prints both the children
and the parents of an item.

Finally, the code in this commit is not of very high quality, so we will want to
clean that up in the not too distant future.