Improve file format
This style should be more extensible. The top-level structure is not a key-value
map rather than a plain list, and it contains a file-verion field which can be
used to ensure backward compatibility in the future.

This commit also cleans up the code for serialization slightly and reduces
coupling between this code and other parts of the program.