Reorganize modules requires
We now require modules relative to the "untask" library rather than as relative
paths.

I'm not really sure how Racket wants you to handle your modules. This approach
is a little better than what we had before, because we no longer need to keep
track of realtive paths, but it does require you to do "raco pkg install" to be
able to do anything.