Add unit tests for evaluating "Today" literals
This unit test pretty much just replicates the functionality of
evaluate-literal, but it may be useful if we want to refactor in the future.

We have to do it this way, since the tests can be run even when the current date
is different. Perhaps there is some way of simulating the current date, but it
doesn't seem worth the effort.