Fix bug where using "color:" as an expression would crash
This was a combination of an issue with how types are handled and the fact that
the type of "color" was set to be an enum rather than an optional enum.

This fix is arguably a bit of a workaround. We will probably need to do an
overhaul of the entire system for checking types and reporting errors at some
point. For now this is good enough.