Improve performance of "agenda" and "schedule" commands
We do this by no longer sorting unnecessarily by urgency.

This fix is essentially a workaround. The problem is that calculating some
properties (in this case the "depends" property) is extremely slow as it
requires looking up every single item in the database for each item that the
property should be calcualted for.

We should change how such properties are caluclated and stored. There is no
reason why looking up a property should not always be O(1).