Fix bugs related to copying calculated properties
The module that manages built-in properties now has a special procedure for
handling cloning.

Eventually, we will probably want to get rid of the old property and item
modules entirely and just have a single module (what is currently
properties/builtin) that can do special behavior like this in a straight-forward
way. Having properties exist as their own separate entities seemed like a good
idea that would decrease coupling, but in fact it has just made it really
difficult to implement certain behavior.