Optimize how link-like properties are stored
Rather than calculating a property like "blocks" every time it is used, we now
store it in the item state and just avoid serializing it to disk.

This makes "list" (and "agenda" and other such commands) much faster when many
items match, since we would previously have to look at every item in the
database for each item in the listing before we could write any output.